requirements.txt
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
pinecone-client
sqlalchemy
pydantic
passlib